Add MySQL and multi-driver DB integration tests
