+++ .github/workflows/tests-plugin-integration.yml
@@ -1,0 +1,72 @@
+name: Plugin Integration Tests
+
+on:
+  pull_request:
+    branches:
+      - main
+  push:
+    branches:
+      - main
+
+concurrency:
+  group: test-plugin-${{ github.head_ref || github.run_id }}
+  cancel-in-progress: true
+
+jobs:
+  plugin-integration:
+    runs-on: ubuntu-latest
+
+    services:
+      redis:
+        image: redis
+        ports:
+          - 6379:6379
+        options: -e REDIS_PASSWORD=difyai123456
+
+      postgres:
+        image: postgres
+        env:
+          POSTGRES_USER: postgres
+          POSTGRES_PASSWORD: difyai123456
+          POSTGRES_DB: dify_plugin_daemon
+        ports:
+          - 5432:5432
+
+      pgbouncer:
+        image: edoburu/pgbouncer
+        env:
+          DB_HOST: postgres
+          DB_PORT: 5432
+          DB_USER: postgres
+          DB_PASSWORD: difyai123456
+          DB_NAME: dify_plugin_daemon
+          POOL_MODE: transaction
+        ports:
+          - 6432:6432
+        options: >-
+          --health-cmd "pg_isready -h localhost -p 6432 -U postgres -d dify_plugin_daemon"
+          --health-interval=10s
+          --health-timeout=5s
+          --health-retries=5
+
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Setup Go
+        uses: actions/setup-go@v5
+        with:
+          go-version: '1.23'
+
+      - name: Install uv
+        run: |
+          curl -LsSf https://astral.sh/uv/install.sh | sh
+          echo "UV_PATH=$(which uv)" >> $GITHUB_ENV
+
+      - name: Setup License
+        run: go run cmd/license/generate/main.go
+
+      - name: Install dependencies
+        run: go mod download
+
+      - name: Run plugin integration tests
+        run: go test -v ./integration/plugin/...
