+++ integration/plugin/testdata/invoke_agent_strategy_json.json
